integerere navbar i felles stil
