+++ css/navigasjon.css
@@ -51,71 +51,4 @@ nav {
   border-right: 0;
 }
 
-.nav-item a {
-  text-decoration: none;
-  color: black;
-}
 
-/* responsiv sidenav */
-
-#nav-list.active {
-  transform: translate3d(0, 0, 0);
-  left: 0;
-}
-
-#nav-list.active+.backdrop {
-  visibility: visible;
-  opacity: 0.4;
-}
-
-#nav-list.touched {
-  transition-duration: 400ms;
-  transition-timing-function: cubic-bezier(0.25, 0.8, 0.25, 1);
-  transition-property: background-color, transform;
-}
-
-.backdrop {
-  background-color: #000;
-  z-index: 999;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  opacity: 0;
-  content: '';
-  position: absolute;
-  visibility: hidden;
-}
-
-.sidenav-button {
-  padding: 0;
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
-  /* lettere å treffe på mobil */
-  background: url("../assets/menu.svg");
-  background-repeat: no-repeat;
-  background-position: center;
-  border: none;
-  align-self: flex-end;
-  display: none;
-}
-
-@media only screen and (max-width: 600px) {
-  #nav-list {
-    transform: translate3d(-100%, 0, 0);
-    flex-direction: column;
-    position: fixed;
-    top: 0;
-    bottom: 0;
-    width: 200px;
-    z-index: 1000;
-    justify-content: initial;
-  }
-  .nav-item {
-    border: 0;
-  }
-  .sidenav-button {
-    display: block;
-  }
-}
