+++ css/style.css
@@ -820,3 +820,118 @@ input[type=submit]:hover {
         box-shadow: inset 0px 0px 0px 80px var(--primary-color);
     }
 }
+
+/* Navigasjonsbar */
+
+.nav-header {
+    padding: 0 1rem;
+}
+
+.nav-header-img {
+    width: 440px;
+    max-width: 100%;
+    margin: auto;
+    display: block;
+}
+
+.navbar {
+    background: var(--primary-color);
+    display: flex;
+    flex-direction: column;
+}
+
+#nav-list {
+    display: flex;
+    list-style: none;
+    flex-direction: row;
+    background: var(--primary-color);
+    margin: 0;
+    padding: .5rem 0;
+    line-height: 1;
+    overflow: hidden;
+    font-size: 1rem;
+    justify-content: center;
+}
+
+.nav-item {
+    border-right: 1px solid rgb(0 0 0 / 50%);
+}
+
+.nav-item a {
+    padding: 8px 16px;
+    display: block;
+    text-decoration: none;
+    color: black;
+}
+
+.nav-item.active {
+    font-weight: bold;
+}
+
+.nav-item:last-of-type {
+    border-right: 0;
+}
+
+/* responsiv sidenav */
+
+#nav-list.active {
+    transform: translate3d(0, 0, 0);
+    left: 0;
+}
+
+#nav-list.active+.backdrop {
+    visibility: visible;
+    opacity: 0.4;
+}
+
+#nav-list.touched {
+    transition-duration: 400ms;
+    transition-timing-function: cubic-bezier(0.25, 0.8, 0.25, 1);
+    transition-property: background-color, transform;
+}
+
+.backdrop {
+    background-color: #000;
+    z-index: 999;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    opacity: 0;
+    content: '';
+    position: absolute;
+    visibility: hidden;
+}
+
+.sidenav-button {
+    padding: 0;
+    width: 40px;
+    height: 40px;
+    border-radius: 50%;
+    /* lettere å treffe på mobil */
+    background: url("../assets/menu.svg");
+    background-repeat: no-repeat;
+    background-position: center;
+    border: none;
+    align-self: flex-end;
+    display: none;
+}
+
+@media only screen and (max-width: 600px) {
+    #nav-list {
+        transform: translate3d(-100%, 0, 0);
+        flex-direction: column;
+        position: fixed;
+        top: 0;
+        bottom: 0;
+        width: 200px;
+        z-index: 1000;
+        justify-content: initial;
+    }
+    .nav-item {
+        border: 0;
+    }
+    .sidenav-button {
+        display: block;
+    }
+}
