+++ index.html
@@ -117,7 +117,8 @@
         </div>
         <fore-footer></fore-footer>
         <script type="module" src="./scripts/hjem.js"></script>
-        <script src="./scripts/nyheter.js"></script>
+        <script type="module" src="./scripts/nyheter.js">
+        </script>
         <script src="./scripts/værmelding.js"></script>
       </body>
     </html>
