+++ scripts/hjem.js
@@ -1,10 +1,14 @@
 import "./common.js";
-class ForeHjemNyhetCard extends NyhetCard {
+import { NyhetBoks, NyhetBoksCollectionElement, Init } from "./nyheter.js";
+
+Init();
+
+class ForeHjemNyhetCard extends NyhetBoks {
   truncateTegn = 221;
 }
 customElements.define("fore-hjem-nyhet", ForeHjemNyhetCard, { extends: "article" });
 
-class ForeHjemNyheterCollection extends NyhetCardCollectionElement {
+class ForeHjemNyheterCollection extends NyhetBoksCollectionElement {
   constructor() {
     super(ForeHjemNyhetCard);
   }
